+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - CI-365 - Changed the "Contact Us" section on the services page in "About DIT" to have its body be markdown
 - CI-368 - Add ebooks promotional section on "why choose the UK" page
 - CI-367 - Update `CapitalInvestRegionPage`
+- CI-366 - Add CTA to UK benefits section of invest home page
 
 
 ### Fixed Bugs